+++ src/users/user.controller.ts
@@ -7,12 +7,11 @@ import {
   Post,
   Put,
   UseGuards,
+  Request,
 } from '@nestjs/common';
 import { plainToInstance } from 'class-transformer';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
-import { PostDTO } from 'src/posts/post.dto';
 import { UserDTO } from './user.dto';
-import { UserEntity } from './user.entity';
 import { UserService } from './user.service';
 
 @Controller('users')
@@ -21,7 +20,12 @@ export class UserController {
 
   @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  async deleteUserById(@Param('id') id: string): Promise<string> {
+  async deleteUserById(
+    @Param('id') id: string,
+    @Request() req,
+  ): Promise<string> {
+    const resultAuthor = await this.userService.authorization(req.user.id, id);
+    if (resultAuthor) return resultAuthor;
     const result = await this.userService.deleteById(id);
     if (result.affected === 1) {
       return 'Delete Successful!';
@@ -32,7 +36,13 @@ export class UserController {
 
   @UseGuards(JwtAuthGuard)
   @Put(':id')
-  async updateUserById(@Param('id') id: string, @Body() user: UserDTO) {
+  async updateUserById(
+    @Param('id') id: string,
+    @Body() user: UserDTO,
+    @Request() req,
+  ) {
+    const result = await this.userService.authorization(req.user.id, id);
+    if (result) return result;
     try {
       await this.userService.updateUserById(id, user);
       return this.getUserById(id);
@@ -41,34 +51,16 @@ export class UserController {
     }
   }
 
-  @Post()
-  createUser(@Body() user: UserDTO): UserDTO {
-    const result = this.userService.saveUser(user);
-    return plainToInstance(UserDTO, result, {
-      excludeExtraneousValues: true,
-    });
-  }
-
   @Get(':id')
   async getUserById(@Param('id') id: string): Promise<UserDTO> {
     const result = await this.userService.getUserById(id);
-    return this.plainUser(result);
+    return this.userService.plainUser(result);
   }
 
   @Get()
   async getAllUsers(): Promise<UserDTO[]> {
     const result = await this.userService.getAllUsers();
-    const plainArray = result.map((res) => this.plainUser(res));
+    const plainArray = result.map((res) => this.userService.plainUser(res));
     return plainArray;
-  }
-
-  plainUser(user: UserEntity): UserDTO {
-    const respl = plainToInstance(UserDTO, user, {
-      excludeExtraneousValues: true,
-    });
-    respl.posts = plainToInstance(PostDTO, user.posts, {
-      excludeExtraneousValues: true,
-    });
-    return respl;
   }
 }
